Initial commit
Jeffrey asked about a tool that would be able to easily sequence the
timestamps on log files, and compute the time intervals - this is a
first cut at that. We are putting in faux data, but we can fix that up
very easily. Still no save/load, but that, too, is possible.