Fix a few things
We didn't want to try and convert 'undefined' in the code, so let's trap
for that, and call it a zero. Also, in the pop-up menu on the table, we
can be a little smarter about the invalidation of adding and removing
rows, and so we have. This is a little bit better.