Fixed up the new parsing
We had a new ask for the timestamp format, and so we had to add a few
things in order to get it all to work. First off, we needed to strip
parts of the timestamp out so that it was parsable, and then add the new
format to the list. Then we had to clean up a little whitespace, and
finally, we had to encode the argument on the call as the new format was
causing the service problems, and this cleaned it right up.