Added the date scanning and skipping
We needed to accept times - without dates, and then scan up the table to
see what the date should be, and use that. At the same time, if no date
is found, assume 'today'. Then, if the times are 'inverted' in the
table, add a day to the lower one to account for an offset.